requirements.txt
google_api_python_client==2.108.0
langdetect==1.0.9
matplotlib==3.6.2
pandas==1.5.2
python-dotenv==1.0.0
requests==2.28.1
textblob==0.17.1
wordcloud==1.9.2
